requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dotenv==0.9.9
duckduckgo_search==8.0.2
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.93.0
google-cloud-bigquery==3.32.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.15.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-google-genai==2.1.4
langchain-google-vertexai==2.0.24
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numexpr==2.10.2
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
validators==0.35.0
websockets==15.0.1
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
